Refactor CountsSourceMixin so it can be included in TabulateSource
