+++ dronefly/core/menus/taxon.py
@@ -1,20 +1,15 @@
 from typing import Any
 
-from pyinaturalist import User
-
 from dronefly.core.clients.inat import iNatClient
-from dronefly.core.formatters.generic import CountsFormatter
-from dronefly.core.menus.counts import CountsSource
-from dronefly.core.query.query import get_place_count, get_user_count
+from dronefly.core.menus.counts import CountsSourceMixin
 
 from .menu import BaseMenu
-from .source import ListPageSource
 from ..formatters import TaxonFormatter
 from ..formatters.constants import WWW_BASE_URL
 from ..query import QueryResponse
 
 
-class TaxonSource(ListPageSource):
+class TaxonSource(CountsSourceMixin):
     def __init__(self, taxon_formatter: TaxonFormatter, with_ancestors: bool = True):
         self._taxon_formatter = taxon_formatter
         self._url = f"{WWW_BASE_URL}/taxon/{self.query_response.taxon.id}"
@@ -42,98 +37,6 @@ class TaxonSource(ListPageSource):
 
     def update_page(self):
         self.entries[0] = self.format_page()
-
-    @property
-    def counts_formatter(self) -> CountsFormatter:
-        return self.formatter.counts_formatter
-
-    @property
-    def counts_page(self) -> str:
-        return self.formatter.counts_page
-
-    @counts_formatter.setter
-    def counts_formatter(self, formatter):
-        self.formatter.counts_formatter = formatter
-
-    @counts_page.setter
-    def counts_page(self, page):
-        self.formatter.counts_page = page
-
-    @property
-    def counts_source(self) -> CountsSource:
-        return self.counts_formatter.source if self.counts_formatter else None
-
-    async def toggle_place_count(self, inat_client: iNatClient, place: User):
-        query_response = self.query_response
-        counts_formatter = self.counts_formatter
-        place_count = None
-        if self.counts_source:
-            place_count = next(
-                (count for count in self.counts_source.entries if count.id == place.id),
-                None,
-            )
-        if place_count is not None:
-            self.counts_source.entries.remove(place_count)
-        else:
-            place_count = await get_place_count(inat_client, query_response, place)
-            if self.counts_source:
-                # A source already exists. Just append the count:
-                self.counts_source.entries.append(place_count)
-            else:
-                # Create both a new source and formatter for counts
-                # with the user count as its only entry and link them
-                # back to the taxon formatter:
-                counts_formatter = CountsFormatter()
-                counts_source = CountsSource(
-                    entries=[place_count],
-                    query_response=query_response,
-                    counts_formatter=counts_formatter,
-                    per_page=15,  # FIXME: magic number!
-                )
-                counts_formatter.source = counts_source
-                self.counts_formatter = counts_formatter
-        # One way or the other, we now have a counts formatter attached. All that remains
-        # is to populate it with new content and regenerate the formatted taxon page
-        # to include it:
-        counts_page = await self.counts_source.get_page(page_number=0)
-        self.formatter.counts_page = counts_page
-        self.update_page()
-
-    async def toggle_user_count(self, inat_client: iNatClient, user: User):
-        query_response = self.query_response
-        counts_formatter = self.counts_formatter
-        user_count = None
-        if self.counts_source:
-            user_count = next(
-                (count for count in self.counts_source.entries if count.id == user.id),
-                None,
-            )
-        if user_count is not None:
-            self.counts_source.entries.remove(user_count)
-        else:
-            user_count = await get_user_count(inat_client, query_response, user)
-            if self.counts_source:
-                # A source already exists. Just append the count:
-                self.counts_source.entries.append(user_count)
-            else:
-                # Create both a new source and formatter for counts
-                # with the user count as its only entry and link them
-                # back to the taxon formatter:
-                counts_formatter = CountsFormatter()
-                counts_source = CountsSource(
-                    entries=[user_count],
-                    query_response=query_response,
-                    counts_formatter=counts_formatter,
-                    per_page=15,  # FIXME: magic number!
-                )
-                counts_formatter.source = counts_source
-                self.counts_formatter = counts_formatter
-        # One way or the other, we now have a counts formatter attached. All that remains
-        # is to populate it with new content and regenerate the formatted taxon page
-        # to include it:
-        counts_page = await self.counts_source.get_page(page_number=0)
-        self.formatter.counts_page = counts_page
-        self.update_page()
 
 
 class TaxonMenu(BaseMenu):
